requirements.txt
pybrowsers==1.1.0
pythondialog==3.5.3
